Avoid colour conversion error with 3+ bands rasters
If more than 3 bands are provided (e.g. presence of an alpha band),  grDevices::convertColor returns an uniformative error "Error in dogamma(x) %*% M : non-conformable arguments".
An alternative could be to convert the 'warning' to a 'stop' (but more informative about what's wrong) and let the user take care of providing a clean 3 bands raster.